Add very basic unit tests and check for coding style
The GitLab CI job "phpcs" is allowed to fail because we do not follow the PSR2 coding style at the moment.
We can fix most of the problems reported there by running `vendor/bin/phpcbf`.
Because this changes all files and nearly all lines,
this should be done in a separate MR.

Related to #27.